Refactor: 各コンポーネントに記述していたバリデーションを1つのファイルにまとめた
vee-validate.jsを作成し記述している。
エラーメッセージのインタラクションモードは各コンポーネントごとに異なるため各々で記述している。
