+++ src/components/GreenLineVCard.vue
@@ -6,12 +6,14 @@
       class="mx-auto"
       elevation="3"
       :max-width="maxWidth"
+      :min-height="minHeight"
     >
       <v-card
         outlined
         color="customAlmostWhite"
         class="rounded-0"
         :max-width="maxWidth"
+        :min-height="minHeight"
       >
         <slot />
       </v-card>
@@ -26,11 +28,16 @@ export default {
     setMaxWidth: {
       type: String,
       default: null
+    },
+    setMinHeight: {
+      type: String,
+      default: null
     }
   },
   data() {
     return {
-      maxWidth: this.setMaxWidth
+      maxWidth: this.setMaxWidth,
+      minHeight: this.setMinHeight
     }
   }
 }
