+++ src/main.js
@@ -4,6 +4,7 @@ import router from './router'
 import store from './store'
 import vuetify from './plugins/vuetify'
 import VueHead from 'vue-head'
+import veeValidate from './plugins/vee-validate'
 
 Vue.use(VueHead)
 
@@ -12,6 +13,7 @@ Vue.config.productionTip = false
 new Vue({
   router,
   vuetify,
+  veeValidate,
   store,
   render: h => h(App)
 }).$mount('#app')
